create useMarkerIcons hook
